Updated eden to use standardized responses
Summary:
Eden previously used WriteJSON with any parameter to respond.
Now, it can still do so, but I gave it a Response struct which
will be the type of response we typically want from our APIs. This
way the programmer can create a Response with the status such as
"ERROR" or "OK" and respond with that instead.